requirements_test.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements_test.txt requirements_test.in
#
-e git+https://github.com/uktrade/directory-components.git@v0.4.1#egg=directory_components
-e git+https://github.com/uktrade/directory-constants.git@v6.4.0#egg=directory_constants
-e git+https://github.com/uktrade/directory-header-footer.git@v2.12.2#egg=directory_header_footer
-e git+https://github.com/uktrade/directory-healthcheck.git@v0.3.0#egg=directory_healthcheck
-e git+https://github.com/uktrade/export-elements.git@v0.3.0#egg=export_elements
-e git+https://github.com/uktrade/directory-signature-auth.git@v2.0.0#egg=sigauth
beautifulsoup4==4.6.0
cachetools==2.0.1         # via google-auth
certifi==2018.1.18        # via requests
chardet==3.0.4            # via requests
codecov==2.0.9
contextlib2==0.5.5        # via raven
coverage==4.4.1           # via codecov, pytest-cov
dj-database-url==0.4.2
django-health-check==3.0.0
django-modelcluster==3.1  # via wagtail
django-modeltranslation==0.12.2  # via wagtail-modeltranslation
django-pglocks==1.0.2
django-taggit==0.22.2     # via wagtail
django-treebeard==4.2.0   # via wagtail
django==1.11.5
djangorestframework==3.4.7  # via wagtail
factory-boy==2.8.1
faker==0.8.11             # via factory-boy
flake8==3.0.4
freezegun==0.3.8
google-api-core==0.1.4    # via google-cloud-core, google-cloud-translate
google-auth==1.4.1        # via google-api-core
google-cloud-core==0.28.0  # via google-cloud-translate
google-cloud-translate==1.3.0
googleapis-common-protos==1.5.3  # via google-api-core
gunicorn==19.5.0
html5lib==0.999999999     # via wagtail
idna==2.6                 # via requests
mccabe==0.5.3             # via flake8
mohawk==0.3.4
olefile==0.45.1           # via pillow
pillow==4.3.0             # via wagtail
protobuf==3.5.1           # via google-api-core, googleapis-common-protos
psycopg2==2.7.3.2
py==1.4.34                # via pytest, pytest-catchlog
pyasn1-modules==0.2.1     # via google-auth
pyasn1==0.4.2             # via pyasn1-modules, rsa
pycodestyle==2.0.0        # via flake8
pyflakes==1.2.3           # via flake8
pytest-catchlog==1.2.2
pytest-cov==2.3.1
pytest-django==3.1.2
pytest-sugar==0.8.0
pytest==3.0.2
python-dateutil==2.6.1    # via faker, freezegun
pytz==2017.2              # via django, django-modelcluster, google-api-core
raven==5.19.0
requests==2.18.4          # via codecov, google-api-core, requests-mock, wagtail
requests_mock==1.1.0
rsa==3.4.2                # via google-auth
setuptools-git==1.2
six==1.11.0               # via faker, freezegun, google-api-core, google-auth, html5lib, mohawk, protobuf, python-dateutil, requests-mock
termcolor==1.1.0          # via pytest-sugar
text-unidecode==1.1       # via faker
unidecode==0.4.21         # via wagtail
urllib3==1.22             # via requests
wagtail-factories==0.3.0
wagtail-modeltranslation==0.8
wagtail==1.13
webencodings==0.5.1       # via html5lib
whitenoise==3.1
willow==1.0               # via wagtail
